FEED_1280: fixed multiple cards issue

See merge request solvers.io/hunt-mobile!587

+++ src/containers/FeedDetailScreen/index.js
@@ -94,7 +94,7 @@ import { TAGS_FEATURE, SHARE_LINK_URL } from "../../service/api"
 
 import Analytics from '../../lib/firebase'
 
-const TOASTER_DURATION = 5000
+const TOASTER_DURATION = 3000
 
 const ACTION_NONE = 0;
 const ACTION_FEEDO_PIN = 1;
@@ -182,6 +182,8 @@ class FeedDetailScreen extends React.Component {
     this.userActions = [];
     this.userActionTimer = null;
     this.scrollviewHeight = 0
+
+    this.deletedCardId = null
   }
 
   UNSAFE_componentWillMount() {
@@ -943,24 +945,32 @@ class FeedDetailScreen extends React.Component {
       return;
     }
     const currentCardInfo = this.userActions[0];
+
     this.setState({ 
       currentActionType: currentCardInfo.currentActionType,
       isShowToaster: true,
       toasterTitle: currentCardInfo.toasterTitle,
     });
+
     this.userActionTimer = setTimeout(() => {
-      // this.setState({ isShowToaster: false })
       if (this.state.currentActionType === ACTION_CARD_DELETE) {
-        Analytics.logEvent('feed_detail_delete_card', {})
-        this.props.deleteCard(currentCardInfo.ideaId)
+        if (this.deletedCardId !== currentCardInfo.ideaId) {
+          Analytics.logEvent('feed_detail_delete_card', {})
+          this.deletedCardId = currentCardInfo.ideaId
+          this.props.deleteCard(currentCardInfo.ideaId)
+          this.userActionTimer = null;
+          this.setState({ isShowToaster: false })
+          this.userActions.shift();
+        }
       } else if (this.state.currentActionType === ACTION_CARD_MOVE) {
         Analytics.logEvent('feed_detail_move_card', {})
         this.props.moveCard(currentCardInfo.ideaId, currentCardInfo.feedoId);
+        this.userActionTimer = null;
+        this.setState({ isShowToaster: false })
+        this.userActions.shift();
       }
-      this.userActionTimer = null;
-      this.userActions.shift();
       this.processCardActions();
-    }, TOASTER_DURATION + 5);
+    }, TOASTER_DURATION + 50);
   }
 
   onDeleteCard = (ideaId) => {
